remove old profiler code
