+++ localstack/utils/bootstrap.py
@@ -9,12 +9,12 @@ import threading
 import time
 import warnings
 from datetime import datetime
+from functools import wraps
 
 import pip as pip_mod
 import six
 
 from localstack import config, constants
-from localstack.utils.analytics.profiler import log_duration
 from localstack.utils.docker import DOCKER_CLIENT, ContainerException, PortMappings
 
 # set up logger
@@ -118,6 +118,29 @@ def run_pip_main(args):
     import pip._internal.main
 
     return pip._internal.main.main(args)
+
+
+def log_duration(name=None):
+    """Function decorator to log the duration of function invocations."""
+
+    def wrapper(f):
+        @wraps(f)
+        def wrapped(*args, **kwargs):
+            from localstack.utils.common import now_utc
+
+            start_time = now_utc(millis=True)
+            try:
+                return f(*args, **kwargs)
+            finally:
+                end_time = now_utc(millis=True)
+                func_name = name or f.__name__
+                duration = end_time - start_time
+                if duration > 500:
+                    LOG.info('Execution of "%s" took %.2fms', func_name, duration)
+
+        return wrapped
+
+    return wrapper
 
 
 @log_duration()
@@ -514,7 +537,6 @@ def extract_port_flags(user_flags, port_mappings):
 
 
 def start_infra_in_docker():
-
     container_name = config.MAIN_CONTAINER_NAME
 
     if DOCKER_CLIENT.is_container_running(container_name):
